Adapt way to parse cli args.
JCommander does not allow String as a Main parameter although it is
written like this in the documentation on their homepage.
In this App we want to specify whether the program should do train,
test or both.

Now the user can specify an only parameter with a string to state, if
he only wants to do one.